pass bridge value to vznetcfg script
currently, only ifacenameis passed to vznetcfg script. (and vznetaddbr)

this is ok for vmstart,
but if we want to change veth value online (new bridge, new vlan, new firewall)
or hotplug a new card, with -netif_add

this is not currently possible,
because we read the bridge from vmid.conf in vznetaddbr, but vmid.conf is updated
after -netif_add (and vznetaddbr).
So we never have new value.

This patch pass bridge value as argument to venetcfg script.
We can now compare new and old value in vznetaddr, and see if it's an update or a new tap create.

Signed-off-by: Alexandre Derumier <aderumier@odiso.com>